Move to separate RA & SA modules
I initially thought that the requesting agency and supplying agency
roles could be combined into a single module. But it has become
increasingly clear that this is just going to lead to confusion down the
line. So I'm splitting them into separate modules.

As it happens, I'd not done anything on the SA role yet, so there's not
anything to split out, I have however rearranged the naming from
ill-storage-ra to ill-ra-storage as I think this will better align with
the API modules, aiming for something like this:

- ill-ra-storage
- ill-sa-storage
- ill-ra
- ill-sa